Lots of improvements to threeworldsdaynightglobe.
Allow the date to advance, to watch the progress of the shadow
across the earth.

Add a new version of OrbitControls called OrbitingOrbitControls
where orbiting actually works, and which doesn't block key bindings
to stop, start and speed up the rotation.

Add said key bindings.

Make the orbit control speed match the speed of the globe's rotation,
so the camera stays hovering over a specific point.

Add a status area where the current date is kept updated.